A new branch to find out lacking of running efficient

+++ core/Reader/Edgeindex.java
@@ -2,10 +2,10 @@ package Reader;
 
 public class Edgeindex {
 	protected int node;
-	protected short count;
+	protected byte count;
 	protected int p2Way;
 
-	public Edgeindex(int node, short count, int p2way) {
+	public Edgeindex(int node, byte count, int p2way) {
 		this.node = node;
 		this.count = count;
 		this.p2Way = p2way;
